improved User, Doctor, Patient and Appointment entities and its use. also added email functionality asynchronously when Admin creates a doctor an email will be sent to doctors email. and doctor can change appointment state from booked to complete (if it is not canceled) and patient can cancel appointment (if it is not completed).
improved User, Doctor, Patient and Appointment entities and its use. also added email functionality asynchronously when Admin creates a doctor an email will be sent to doctors email. and doctor can change appointment state from booked to complete (if it is not canceled) and patient can cancel appointment (if it is not completed).